Reworked drawing
Note: very broken but bug tested... I think.

+++ src/io/shapez/managers/providers/MiscProvider.java
@@ -50,7 +50,7 @@ public class MiscProvider {
         return splashes[rng.nextInt(splashes.length)];
     }
 
-    public static float clamp(float val, float min, float max) {
+    public static double clamp(double val, double min, double max) {
         return Math.max(min, Math.min(max, val));
     }
